proxmox/python-tools/requirements.txt
requests>=2.31.0
pyyaml>=6.0
urllib3>=2.0.0
python-dotenv>=1.0.0

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
mypy>=1.0.0
types-requests>=2.31.0
pydantic>=2.0.0

